requirements.txt
fastapi==0.88.0
pydantic~=1.10.2
passlib~=1.7.4
SQLAlchemy~=1.4.45
PyJWT~=2.6.0
python-multipart~=0.0.5
bcrypt~=4.0.1